+++ src/formatters/checkstyleFormatter.ts
@@ -17,7 +17,7 @@
 
 import {AbstractFormatter} from "../language/formatter/abstractFormatter";
 import {IFormatterMetadata} from "../language/formatter/formatter";
-import {RuleFailure, RuleLevel} from "../language/rule/rule";
+import {RuleFailure, RuleSeverity} from "../language/rule/rule";
 
 import * as Utils from "../utils";
 
@@ -48,7 +48,7 @@ export class Formatter extends AbstractFormatter {
             });
             let previousFilename: string | null = null;
             for (const failure of failuresSorted) {
-                const level = RuleLevel[failure.getRuleLevel()].toLowerCase();
+                const severity = RuleSeverity[failure.getRuleSeverity()].toLowerCase();
                 if (failure.getFileName() !== previousFilename) {
                     if (previousFilename) {
                         output += "</file>";
@@ -58,7 +58,7 @@ export class Formatter extends AbstractFormatter {
                 }
                 output += "<error line=\"" + (failure.getStartPosition().getLineAndCharacter().line + 1) + "\" ";
                 output += "column=\"" + (failure.getStartPosition().getLineAndCharacter().character + 1) + "\" ";
-                output += "severity=\"" + level + "\" ";
+                output += "severity=\"" + severity + "\" ";
                 output += "message=\"" + this.escapeXml(failure.getFailure()) + "\" ";
                 // checkstyle parser wants "source" to have structure like <anything>dot<category>dot<type>
                 output += "source=\"failure.tslint." + this.escapeXml(failure.getRuleName()) + "\" />";